+++ R/package.R
@@ -45,3 +45,28 @@
 #'
 #' @import sf
 NULL
+
+#' Verbose API usage of hereR
+#'
+#' If set to \code{TRUE} the hereR package is messaging information about
+#' the amount of requests sent to the APIs and data size received.
+#'
+#' @param ans boolean, verbose or not (default = \code{FALSE})?
+#'
+#' @return
+#' None.
+#'
+#' @export
+#'
+#' @examples
+#' set_verbose(TRUE)
+set_verbose <- function(ans = FALSE) {
+  .check_boolean(ans)
+  if (ans) {
+    Sys.setenv(
+      "HERE_VERBOSE" = "TRUE"
+    )
+  } else {
+    Sys.unsetenv("HERE_VERBOSE")
+  }
+}